@@ -1,7 +1,7 @@
 /**
  *  \file category.hpp
  *
- *  Copyright 2014 Michael Caisse : ciere.com
+ *  Copyright 2014, 2015 Michael Caisse : ciere.com
  *
  *  Distributed under the Boost Software License, Version 1.0. (See accompanying
  *  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
@@ -17,7 +17,7 @@ namespace boost { namespace boostache { namespace extension
    struct category_attribute {};
    struct unused_attribute : category_attribute {};
    struct plain_attribute : category_attribute {};
-   struct container_attribute : category_attribute {};
+   struct sequence_attribute : category_attribute {};
    struct associative_attribute : category_attribute {};
    struct tuple_attribute : category_attribute {};
    struct variant_attribute : category_attribute {};
