@@ -27,19 +27,15 @@ namespace boost { namespace boostache { namespace extension
 
    template <typename T>
    struct select_category< T
-                         , typename enable_if<vm::trait::is_variant<T>>::type>
+                         , vm::trait::enable_if_is_variant_t<T>
+                         >
       : mpl::identity<variant_attribute> {};
 
    template <typename T>
    struct select_category< T
-                         , typename std::enable_if<
-                              vm::trait::has_begin<
-                                 typename vm::trait::not_a_map<T>::type 
-                                 >::value 
-                              >::type
+                         , vm::trait::enable_if_sequence_not_map_t<T>
                          >
-      : mpl::identity<container_attribute>
-   {};
+      : mpl::identity<sequence_attribute> {};
 
    template <typename T>
    struct select_category<boost::optional<T>>
@@ -48,6 +44,9 @@ namespace boost { namespace boostache { namespace extension
    template <typename T>
    struct select_category<std::map<std::string,T>>
       : mpl::identity<associative_attribute> {};
+
+   template <typename T>
+   using select_category_t = typename select_category<T>::type;
 
 }}}
 
