@@ -13,8 +13,8 @@
  *  Distributed under the Boost Software License, Version 1.0. (See accompanying
  *  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
  */
-//#define BOOST_TEST_DYN_LINK
 #define BOOST_TEST_NO_MAIN
+#define BOOST_TEST_ALTERNATIVE_INIT_API
 
 #include <boost/filesystem.hpp>
 #include <boost/test/unit_test.hpp>
